update settings page layout
